# projecteBackend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Conecta Salud API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/calls": {
            "get": {
                "tags": [
                    "Calls"
                ],
                "summary": "Obté totes les trucades",
                "operationId": "9a23d14f999134958731dd0346155df0",
                "responses": {
                    "200": {
                        "description": "Llista de trucades obtinguda correctament",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CallResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Calls"
                ],
                "summary": "Crea una nova trucada",
                "operationId": "db4e93d17f5ef9ad4a6f3a161cf30a45",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreCallRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Trucada creada correctament",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CallResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error de validació o falta d'informació"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/calls/{id}": {
            "get": {
                "tags": [
                    "Calls"
                ],
                "summary": "Obté una trucada per ID",
                "operationId": "e72c3e9ae2eed11b9dce29f58032517b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la trucada",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 3
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Trucada trobada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CallResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Trucada no trobada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Calls"
                ],
                "summary": "Actualitza una trucada existent",
                "operationId": "0104abb19faaa5bca335c22b0762e2d6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la trucada a actualitzar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 3
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCallRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Trucada actualitzada correctament",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CallResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error en actualitzar la trucada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Calls"
                ],
                "summary": "Elimina una trucada",
                "operationId": "7ea28db82b97c7b3fe159f3c394be63d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la trucada a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 3
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Trucada eliminada correctament"
                    },
                    "404": {
                        "description": "Trucada no trobada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreCallRequest": {
                "description": "Validació per a l'emmagatzematge de trucades",
                "required": [
                    "patientId",
                    "operatorId",
                    "details",
                    "dateTime"
                ],
                "properties": {
                    "patientId": {
                        "description": "ID del pacient",
                        "type": "integer",
                        "example": 1
                    },
                    "operatorId": {
                        "description": "Correu electrònic únic de l'usuari",
                        "type": "integer",
                        "format": "email",
                        "maxLength": 255,
                        "example": "toni.garcia@example.com"
                    }
                },
                "type": "object"
            },
...
            "UpdateCallRequest": {
                "description": "Validació per a l'actualització de trucades",
                "required": [
                    "patientId",
                    "operatorId",
                    "details",
                    "dateTime"
                ],
                "properties": {
                    "patientId": {
                        "description": "ID del pacient",
                        "type": "integer",
                        "example": 1
                    },
                    "operatorId": {
                        "description": "ID de l'operador",
                        "type": "integer",
                        "example": 2
                    },
                    "details": {
                        "description": "Detalls de la trucada",
                        "type": "string",
                        "example": "Detalls de la trucada"
                    },
                    "dateTime": {
                        "description": "Data i hora de la trucada",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-10-01 12:00:00"
                    },
                    "incomingCall": {
                        "description": "Detalls de la trucada entrant",
                        "properties": {
                            "type": {
                                "description": "Tipus de trucada entrant",
                                "type": "string",
                                "example": "emergency"
                            },
                            "emergencyLevel": {
                                "description": "Nivell d'emergència de la trucada entrant",
                                "type": "integer",
                                "example": 3
                            }
                        },
                        "type": "object"
                    },
                    "outgoingCall": {
                        "description": "Detalls de la trucada sortint",
                        "properties": {
                            "type": {
                                "description": "Tipus de trucada sortint",
                                "type": "string",
                                "example": "follow-up"
                            },
                            "alertId": {
                                "description": "ID de l'alerta associada",
                                "type": "integer",
                                "example": 5
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "CallResource": {
                "description": "Esquema del recurs Call",
                "properties": {
                    "id": {
                        "description": "Identificador de la trucada",
                        "type": "integer",
                        "example": 1
                    },
                    "details": {
                        "description": "Detalls de la trucada",
                        "type": "string",
                        "example": "Detalls de la trucada"
                    },
                    "dateTime": {
                        "description": "Data i hora de la trucada",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-02-20T14:30:00Z"
                    },
                    "operator": {
                        "$ref": "#/components/schemas/OperatorResource"
                    },
                    "patient": {
                        "$ref": "#/components/schemas/PatientSimpleResource"
                    },
                    "incomingCall": {
                        "$ref": "#/components/schemas/IncomingCallResource"
                    },
                    "outgoingCall": {
                        "$ref": "#/components/schemas/OutgoingCallResource"
                    }
                },
                "type": "object"
            },
            "ContactPersonResource": {
                "description": "Esquema del recurs ContactPerson",
                "properties": {
                    "id": {
                        "description": "Identificador de la persona de contacte",
                        "type": "integer",
                        "example": 1
                    },
                    "firstName": {
                        "description": "Nom de la persona de contacte",
                        "type": "string",
                        "example": "John"
                    },
                    "lastName": {
                        "description": "Cognom de la persona de contacte",
                        "type": "string",
                        "example": "Doe"
                    },
                    "phone": {
                        "description": "Telèfon de la persona de contacte",
                        "type": "string",
                        "example": "+123456789"
                    },
                    "relationship": {
                        "description": "Relació amb el pacient",
                        "type": "string",
                        "example": "Germà"
                    },
                    "patient": {
                        "$ref": "#/components/schemas/PatientResource"
                    }
                },
                "type": "object"
            },
            "IncomingCallResource": {
                "description": "Esquema del recurs IncomingCall",
                "properties": {
                    "callId": {
                        "description": "Identificador de la trucada",
                        "type": "integer",
                        "example": 1
                    },
                    "type": {
                        "description": "Tipus de trucada",
                        "type": "string",
                        "example": "emergència"
                    },
                    "emergencyLevel": {
                        "description": "Nivell d'emergència",
                        "type": "integer",
                        "example": 3
                    }
                },
                "type": "object"
            },
            "LanguageResource": {
                "description": "Esquema del recurs Language",
                "properties": {
                    "id": {
                        "description": "Identificador de l'idioma",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Nom de l'idioma",
                        "type": "string",
                        "example": "Català"
                    }
                },
                "type": "object"
            },
            "OperatorResource": {
                "description": "Esquema del recurs Operator",
                "properties": {
                    "id": {
                        "description": "Identificador de l'operador",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Nom de l'operador",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "description": "Correu electrònic de l'operador",
                        "type": "string",
                        "example": "john.doe@example.com"
                    },
                    "role": {
                        "description": "Rol de l'operador",
                        "type": "string",
                        "example": "admin"
                    },
                    "surnames": {
                        "description": "Cognoms de l'operador",
                        "type": "string",
                        "example": "Doe"
                    },
                    "phone": {
                        "description": "Telèfon de l'operador",
                        "type": "string",
                        "example": "123456789"
                    },
                    "hireDate": {
                        "description": "Data de contractació de l'operador",
                        "type": "string",
                        "format": "date",
                        "example": "2022-01-01"
                    },
                    "terminationDate": {
                        "description": "Data de finalització de l'operador",
                        "type": "string",
                        "format": "date",
                        "example": "2023-01-01",
                        "nullable": true
                    },
                    "zones": {
                        "description": "Zones associades a l'operador",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ZoneResource"
                        }
                    },
                    "patients": {
                        "description": "Pacients associats a l'operador",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PatientResource"
                        }
                    },
                    "languages": {
                        "description": "Idiomes parlats per l'operador",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/LanguageResource"
                        }
                    }
                },
                "type": "object"
            },
            "OutgoingCallResource": {
                "description": "Esquema del recurs OutgoingCall",
                "properties": {
                    "callId": {
                        "description": "Identificador de la trucada",
                        "type": "integer",
                        "example": 1
                    },
                    "type": {
                        "description": "Tipus de trucada",
                        "type": "string",
                        "example": "emergència"
                    },
                    "alert": {
                        "$ref": "#/components/schemas/AlertResource"
                    }
                },
                "type": "object"
            },
            "PatientSimpleResource": {
                "description": "Informació simplificada del pacient",
                "properties": {
                    "id": {
                        "description": "Identificador del pacient",
                        "type": "integer",
                        "example": 12
                    },
                    "fullName": {
                        "description": "Nom complet del pacient",
                        "type": "string",
                        "example": "Ana García López"
                    },
                    "birthDate": {
                        "description": "Data de naixement",
                        "type": "string",
                        "format": "date",
                        "example": "1985-07-22"
                    },
                    "fullAddress": {
                        "description": "Adreça completa",
                        "type": "string",
                        "example": "Carrer Major, 45, 2n A, Alcoi, 03801"
                    },
                    "dni": {
                        "description": "DNI del pacient",
                        "type": "string",
                        "example": "12345678A"
                    },
                    "healthCardNumber": {
                        "description": "Número de la targeta sanitària",
                        "type": "string",
                        "example": "SIP87654321"
                    },
                    "phone": {
                        "description": "Número de telèfon",
                        "type": "string",
                        "example": "+34 654 321 987"
                    },
                    "email": {
                        "description": "Correu electrònic",
                        "type": "string",
                        "format": "email",
                        "example": "ana.garcia@example.com"
                    },
                    "personalFamilySituation": {
                        "description": "Situació personal i familiar",
                        "type": "string",
                        "example": "Viu sola, té dos fills que viuen a prop."
                    },
                    "healthSituation": {
                        "description": "Situació de salut",
                        "type": "string",
                        "example": "Diabetis tipus 2 controlada amb medicació."
                    },
                    "housingSituation": {
                        "description": "Situació d'habitatge",
                        "type": "string",
                        "example": "Habitatge propi amb ascensor."
                    },
                    "personalAutonomy": {
                        "description": "Autonomia personal",
                        "type": "string",
                        "example": "Autònoma per a activitats bàsiques, necessita ajuda per a la compra."
                    },
                    "economicSituation": {
                        "description": "Situació econòmica",
                        "type": "string",
                        "example": "Pensió contributiva de 800€ mensuals."
                    }
                },
                "type": "object"
            },
            "PatientResource": {
                "description": "Esquema del recurs Patient",
                "properties": {
                    "id": {
                        "description": "Identificador del pacient",
                        "type": "integer",
                        "example": 1
                    },
                    "fullName": {
                        "description": "Nom complet del pacient",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "birthDate": {
                        "description": "Data de naixement del pacient",
                        "type": "string",
                        "format": "date",
                        "example": "1980-01-01"
                    },
                    "fullAddress": {
                        "description": "Adreça completa del pacient",
                        "type": "string",
                        "example": "123 Main St, City, Country"
                    },
                    "dni": {
                        "description": "DNI del pacient",
                        "type": "string",
                        "example": "12345678A"
                    },
                    "healthCardNumber": {
                        "description": "Número de targeta sanitària del pacient",
                        "type": "string",
                        "example": "HC123456789"
                    },
                    "phone": {
                        "description": "Telèfon del pacient",
                        "type": "string",
                        "example": "+123456789"
                    },
                    "email": {
                        "description": "Correu electrònic del pacient",
                        "type": "string",
                        "format": "email",
                        "example": "johndoe@example.com"
                    },
                    "personalFamilySituation": {
                        "description": "Situació familiar personal del pacient",
                        "type": "string",
                        "example": "Single"
                    },
                    "healthSituation": {
                        "description": "Situació de salut del pacient",
                        "type": "string",
                        "example": "Healthy"
                    },
                    "housingSituation": {
                        "description": "Situació d'habitatge del pacient",
                        "type": "string",
                        "example": "Own house"
                    },
                    "personalAutonomy": {
                        "description": "Autonomia personal del pacient",
                        "type": "string",
                        "example": "Independent"
                    },
                    "economicSituation": {
                        "description": "Situació econòmica del pacient",
                        "type": "string",
                        "example": "Stable"
                    },
                    "operatorId": {
                        "description": "Identificador de l'operador",
                        "type": "integer",
                        "example": 2
                    },
                    "languages": {
                        "description": "Llengües parlades pel pacient",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/LanguageResource"
                        }
                    },
                    "contactPersons": {
                        "description": "Persones de contacte del pacient",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ContactPersonResource"
                        }
                    },
                    "zone": {
                        "$ref": "#/components/schemas/ZoneResource"
                    },
                    "operator": {
                        "$ref": "#/components/schemas/OperatorResource"
                    },
                    "calls": {
                        "description": "Trucades associades al pacient",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CallResource"
                        }
                    }
                },
                "type": "object"
            },
            "ZoneResource": {
                "description": "Esquema del recurs Zone",
                "properties": {
                    "id": {
                        "description": "Identificador de la zona",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Nom de la zona",
                        "type": "string",
                        "example": "Zona 1"
                    },
                    "status": {
                        "description": "Estat de la zona",
                        "type": "string",
                        "example": "activa"
                    },
                    "created_at": {
                        "description": "Data de creació",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T00:00:00Z"
                    },
                    "updated_at": {
                        "description": "Data d'actualització",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-02T00:00:00Z"
                    },
                    "operators": {
                        "description": "Llista d'operadors associats a la zona",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OperatorResource"
                        }
                    },
                    "patients": {
                        "description": "Llista de pacients associats a la zona",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PatientResource"
                        }
                    },
                    "alerts": {
                        "description": "Llista d'alertes associades a la zona",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AlertResource"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}